bugfixing after first test

+++ src/4BS_Telegram.h
@@ -2,7 +2,7 @@
 
 #include "EnoceanProfils.h"
 
-uint8_t handle_4BS(PACKET_SERIAL_TYPE *f_Pkt_st, uint8_t profil, uint8_t profil2nd, uint8_t firstComObj, uint8_t firstParameter)
+uint8_t handle_4BS(PACKET_SERIAL_TYPE *f_Pkt_st, uint8_t profil, uint8_t profil2nd, uint16_t firstComObj, uint16_t firstParameter)
 {
 
       uint16_t lux;
